services/face_recog/requirements.txt
fastapi
uvicorn
numpy
opencv-python
opencv-contrib-python
sqlmodel
python-multipart
pydantic
Pillow
scipy
scikit-image
requests
